// packages/zod-schemas/src/generated/inputTypeSchemas/UserWhereInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { IntFilterSchema } from './IntFilterSchema';
import { StringFilterSchema } from './StringFilterSchema';
import { StringNullableFilterSchema } from './StringNullableFilterSchema';
import { DateTimeFilterSchema } from './DateTimeFilterSchema';
import { CommentListRelationFilterSchema } from './CommentListRelationFilterSchema';
import { SocialTokensNullableScalarRelationFilterSchema } from './SocialTokensNullableScalarRelationFilterSchema';
import { SocialTokensWhereInputSchema } from './SocialTokensWhereInputSchema';

export const UserWhereInputSchema: z.ZodType<Prisma.UserWhereInput> = z.object({
  AND: z.union([ z.lazy(() => UserWhereInputSchema),z.lazy(() => UserWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => UserWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => UserWhereInputSchema),z.lazy(() => UserWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => IntFilterSchema),z.number() ]).optional(),
  email: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  name: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  password: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  refreshToken: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  comments: z.lazy(() => CommentListRelationFilterSchema).optional(),
  socialTokens: z.union([ z.lazy(() => SocialTokensNullableScalarRelationFilterSchema),z.lazy(() => SocialTokensWhereInputSchema) ]).optional().nullable(),
}).strict();

export default UserWhereInputSchema;